refactor: remove test doctype from core
This was commited by mistake, instead of using the test doctype, create
it on demand for tests.
