--- frappe/core/doctype/test/test.js
@@ -1,8 +1,0 @@
-// Copyright (c) 2021, Frappe Technologies and contributors
-// For license information, please see license.txt
-
-frappe.ui.form.on('test', {
-	// refresh: function(frm) {
-
-	// }
-});
